requirements.txt
fvcore==0.1.5.post20221221
numpy==1.26.2
omegaconf==2.3.0
opencv-python==4.8.1.78
opencv-python-headless==4.6.0.66
pandas==1.4.4
Pillow==10.3.0
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.11.4
timm==0.9.12
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.3
xformers==0.0.21
yacs==0.1.8
